refactor: restructure project modules and enhance HTTP response handling
- Moved `api` module to `http` to better reflect purpose.
- Consolidated response and problem types under `http`.
- Refactored pagination logic and moved into dedicated `pagination` module.
- Introduced `contract` module for shared traits like `PersistenceClient`.
- Adjusted imports and internal module visibility accordingly.
- Updated documentation to outline project structure and usage.
