--- src/api.rs
@@ -1,1 +1,0 @@
-pub mod response;
