+++ src/auth/contract/user_details_finder.rs
@@ -1,5 +1,5 @@
-use crate::api::response::ApiError;
 use crate::auth::contract::UserDetails;
+use crate::http::ApiError;
 
 pub trait UserDetailsFinder<TUserDetails>
 where
