+++ src/command/mod.rs
@@ -1,5 +1,5 @@
-pub mod executable;
-pub mod handler;
+mod executable;
+mod handler;
 
 pub use executable::*;
 pub use handler::*;
