+++ src/contract/mod.rs
@@ -1,0 +1,3 @@
+mod persistence_client;
+
+pub use persistence_client::*;
