+++ src/contract/persistence_client.rs
@@ -1,0 +1,1 @@
+pub trait PersistenceClient: Send + Sync + 'static {}
