+++ src/http/response/bad_request.rs
@@ -1,7 +1,7 @@
-use crate::types::problem::{Field, ProblemDetails};
-use axum::Json;
+use crate::http::{Field, ProblemDetails};
 use axum::http::StatusCode;
 use axum::response::{IntoResponse, Response};
+use axum::Json;
 use std::borrow::Cow;
 use validator::ValidationErrors;
