+++ src/http/response/internal_server.rs
@@ -1,7 +1,7 @@
-use crate::types::problem::ProblemDetails;
-use axum::Json;
+use crate::http::ProblemDetails;
 use axum::http::StatusCode;
 use axum::response::{IntoResponse, Response};
+use axum::Json;
 
 pub struct InternalServer;
