+++ src/http/types/problem_details.rs
@@ -1,26 +1,10 @@
+use crate::http::types::Field;
 use serde::Serialize;
 
 #[derive(Serialize)]
 pub struct ProblemDetails {
     pub detail: String,
     pub errors: Vec<Field>,
-}
-
-#[derive(Serialize)]
-pub struct Field {
-    pub field: String,
-    pub reason: String,
-    pub code: String,
-}
-
-impl Field {
-    pub fn new(field: &str, reason: &str, code: &str) -> Self {
-        Self {
-            field: field.to_string(),
-            reason: reason.to_string(),
-            code: code.to_string(),
-        }
-    }
 }
 
 impl From<String> for ProblemDetails {
