+++ src/lib.rs
@@ -1,7 +1,8 @@
-pub mod api;
 #[cfg(feature = "jwt-auth")]
 pub mod auth;
 pub mod command;
+pub mod contract;
+pub mod http;
 #[cfg(feature = "orm")]
 pub mod orm;
-pub mod types;
+pub mod pagination;
