--- src/orm.rs
@@ -1,4 +1,0 @@
-mod client;
-pub mod paginator;
-
-pub use paginator::*;
