+++ src/orm/client.rs
@@ -1,4 +1,4 @@
-use crate::command::PersistenceClient;
+use crate::contract::PersistenceClient;
 
 impl PersistenceClient for sea_orm::DatabaseConnection {}
 impl PersistenceClient for sea_orm::DatabaseTransaction {}
