+++ src/pagination/paginator/contract/paginate.rs
@@ -1,0 +1,14 @@
+use crate::pagination::{Paged, Pagination};
+use sea_orm::{ConnectionTrait, DbErr, SelectorTrait};
+
+pub trait Paginate<'db, C, S>
+where
+    C: ConnectionTrait,
+    S: SelectorTrait,
+{
+    fn paginate(
+        self,
+        conn: &'db C,
+        pagination: &Pagination,
+    ) -> impl Future<Output = Result<Paged<S::Item>, DbErr>>;
+}
