+++ src/pagination/paginator/mod.rs
@@ -1,0 +1,4 @@
+mod contract;
+mod paginate;
+
+pub use contract::*;
