+++ src/pagination/types/search_query.rs
