--- src/types.rs
@@ -1,3 +1,0 @@
-pub mod filter;
-pub mod pagination;
-pub mod problem;
